raml1.0-rc2 fix for resource, method, body, response, and security
