+++ codegen/method.go
@@ -118,10 +118,10 @@ func newClientMethod(r *raml.Resource, rd *resourceDef, m *raml.Method, methodNa
 	}
 }
 
-// assignBodyName assign method's request body by bodies.Type or bodies.ApplicationJson
+// assignBodyName assign method's request body by bodies.Type or bodies.ApplicationJSON
 // if bodiesType generated from bodies.Type we dont need append prefix and suffix
 // 		example : bodies.Type = City, so bodiesType = City
-// if bodiesType generated from bodies.ApplicationJson, we get that value from prefix and suffix
+// if bodiesType generated from bodies.ApplicationJSON, we get that value from prefix and suffix
 //		suffix = [ReqBody | RespBody] and prefix should be uri + method name.
 //		example prefix could be UsersUserIdDelete
 func assignBodyName(bodies raml.Bodies, prefix, suffix string) string {
@@ -129,9 +129,9 @@ func assignBodyName(bodies raml.Bodies, prefix, suffix string) string {
 
 	if len(bodies.Type) > 0 {
 		bodiesType = convertToGoType(bodies.Type)
-	} else if bodies.ApplicationJson != nil {
-		if bodies.ApplicationJson.Type != "" {
-			bodiesType = convertToGoType(bodies.ApplicationJson.Type)
+	} else if bodies.ApplicationJSON != nil {
+		if bodies.ApplicationJSON.Type != "" {
+			bodiesType = convertToGoType(bodies.ApplicationJSON.Type)
 		} else {
 			bodiesType = prefix + suffix
 		}
